combine getImage and upload image route, refacotr getImage function and clean up.
