+++ server/controllers/image.js
@@ -1,5 +1,5 @@
 const asyncHandler = require("express-async-handler");
-const { uploadFile } = require('../utils/s3');
+const { uploadFile, getFileStream } = require('../utils/s3');
 const fs = require('fs');
 const util = require('util');
 const unlinkFile = util.promisify(fs.unlink);
@@ -9,13 +9,17 @@ const unlinkFile = util.promisify(fs.unlink);
 // @desc Upload image to S3
 // @access Private
 exports.uploadImage = asyncHandler(async (req, res, next) => {
-  try {
     const file = req.file;
     const result = await uploadFile(file);
     await unlinkFile(file.path);
-    res.status(200).json({ imagePath: `/images/${result.Key}`});
-  }
-  catch (err) {
-    res.status(500).json({ error: err });
-  }
+    res.status(200).json({ key: result.Key});
+});
+
+// @route GET /images/:key
+// @desc create read stream for image from S3 bucket
+// @access Private
+exports.getImage = asyncHandler(async (req, res, next) => {
+  const imageKey = req.params.key
+  const readStream = await getFileStream(imageKey);
+  res.send(readStream);
 });
