--- server/models/Request.js
@@ -1,33 +1,0 @@
-const mongoose = require("mongoose");
-
-const requestSchema = new mongoose.Schema({
-  userId: {
-    type: String,
-    required: true
-  },
-  sitterId: {
-    type: String,
-    required: true
-  },
-  start: {
-    type: Date,
-    required: true,
-    default: Date.now
-  },
-  end: {
-    type: Date,
-    required: true,
-    default: Date.now
-  },
-  status: {
-    type: String,
-    enum: ["accepted", "declined", "pending"],
-    default: "pending"
-  },
-  paid: {
-    type: Boolean,
-    default: false
-  }
-});
-
-module.exports = Request = mongoose.model("request", requestSchema);
