--- server/routes/uploadImage.js
@@ -1,11 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const multer = require('multer');
-const upload = multer({ dest: 'uploads/' })
-const { uploadImage } = require("../controllers/uploadImage");
-
-
-router.route("/").post(upload.single('image'), uploadImage);
-
-module.exports = router;
-
